extracted test code intro trait
