+++ layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Integration/QueryableCustomPostsOnSchemaConfigurationCPTBlockAttributesFixtureEndpointWebserverRequestTest.php
@@ -10,6 +10,8 @@ use PHPUnitForGraphQLAPI\GraphQLAPI\Integration\AbstractModifyCPTBlockAttributes
 
 class QueryableCustomPostsOnSchemaConfigurationCPTBlockAttributesFixtureEndpointWebserverRequestTest extends AbstractModifyCPTBlockAttributesFixtureEndpointWebserverRequestTestCase
 {
+    use QueryableCustomPostsFixtureEndpointWebserverRequestTestTrait;
+
     public const MOBILE_APP_SCHEMA_CONFIGURATION_ID = 193;
 
     protected function getEndpoint(): string
@@ -22,42 +24,14 @@ class QueryableCustomPostsOnSchemaConfigurationCPTBlockAttributesFixtureEndpoint
         return 'graphql/mobile-app/';
     }
 
-    protected function getFixtureFolder(): string
-    {
-        return __DIR__ . '/fixture-queryable-customposts';
-    }
-
     /**
      * @return array<string,mixed>
      */
     protected function getCPTBlockAttributesNewValue(): array
     {
-        $value = [
-            'post',
-            'attachment',
-            'nav_menu_item',
-            'custom_css',
-            'revision',
-        ];
-
-        $dataName = $this->getDataName();
-        if (str_ends_with($dataName, ':1')) {
-            $value[] = 'page';
-        } elseif (str_ends_with($dataName, ':2')) {
-            $value[] = 'page';
-            $value[] = 'dummy-cpt';
-        }
-
-        /**
-         * Sort them as to store the entries in same way as via the UI,
-         * then tests won't fail whether data was added via PHPUnit test or
-         * via user interface
-         */
-        sort($value);
-
         return [
             AbstractSchemaConfigCustomizableConfigurationBlock::ATTRIBUTE_NAME_CUSTOMIZE_CONFIGURATION => 'true',
-            SchemaConfigSchemaCustomPostsBlock::ATTRIBUTE_NAME_INCLUDED_CUSTOM_POST_TYPES => $value,
+            SchemaConfigSchemaCustomPostsBlock::ATTRIBUTE_NAME_INCLUDED_CUSTOM_POST_TYPES => $this->getIncludedCustomPostTypesNewValue(),
         ];
     }
